muffin/requirements.txt
muffin==0.0.99
muffin-peewee==0.0.43
